moved allowUrls and denyUrls from dart to flutter
