+++ dart/lib/src/sentry_options.dart
@@ -184,21 +184,6 @@ class SentryOptions {
   /// sent. Events are picked randomly. Default is null (disabled)
   double? sampleRate;
 
-  /// (Web only) Events only occurring on these Urls will be handled and sent to sentry.
-  /// If an empty list is used, the SDK will send all errors.
-  /// To use regex add the `^` and the `$` to the string.
-  ///
-  /// If used on a platform other than Web, this setting will be ignored.
-  List<String> allowUrls = [];
-
-  /// (Web only) Events occurring on these Urls will be ignored and are not sent to sentry.
-  /// If an empty list is used, the SDK will send all errors.
-  /// In combination with `allowUrls` you can block subdomains of the domains listed in `allowUrls`.
-  /// To use regex add the `^` and the `$` to the string.
-  ///
-  /// If used on a platform other than Web, this setting will be ignored.
-  List<String> denyUrls = [];
-
   /// The ignoreErrors tells the SDK which errors should be not sent to the sentry server.
   /// If an null or an empty list is used, the SDK will send all transactions.
   /// To use regex add the `^` and the `$` to the string.
