+++ flutter/lib/src/event_processor/url_filter/url_filter_event_processor.dart
@@ -1,8 +1,8 @@
-import '../../../sentry.dart';
+import '../../../sentry_flutter.dart';
 import 'io_url_filter_event_processor.dart'
     if (dart.library.html) 'web_url_filter_event_processor.dart';
 
 abstract class UrlFilterEventProcessor implements EventProcessor {
-  factory UrlFilterEventProcessor(SentryOptions options) =>
+  factory UrlFilterEventProcessor(SentryFlutterOptions options) =>
       urlFilterEventProcessor(options);
 }
