+++ flutter/lib/src/event_processor/url_filter/web_url_filter_event_processor.dart
@@ -2,11 +2,14 @@
 // ignore: depend_on_referenced_packages
 import 'package:web/web.dart' as web show window, Window;
 
-import '../../../sentry.dart';
-import '../../utils/regex_utils.dart';
+import '../../../sentry_flutter.dart';
 import 'url_filter_event_processor.dart';
+// ignore: implementation_imports
+import 'package:sentry/src//utils/regex_utils.dart';
 
-UrlFilterEventProcessor urlFilterEventProcessor(SentryOptions options) =>
+// ignore_for_file: invalid_use_of_internal_member
+
+UrlFilterEventProcessor urlFilterEventProcessor(SentryFlutterOptions options) =>
     WebUrlFilterEventProcessor(options);
 
 class WebUrlFilterEventProcessor implements UrlFilterEventProcessor {
@@ -16,7 +19,7 @@ class WebUrlFilterEventProcessor implements UrlFilterEventProcessor {
 
   final web.Window _window = web.window;
 
-  final SentryOptions _options;
+  final SentryFlutterOptions _options;
 
   @override
   SentryEvent? apply(SentryEvent event, Hint hint) {
